requirements.txt
asgiref==3.5.0
autopep8==1.6.0
Django==4.0.2
django-ckeditor==6.2.0
django-dotenv==1.4.2
django-js-asset==2.0.0
django-quill-editor==0.1.40
django-tinymce==3.4.0
gunicorn==20.1.0
Pillow==9.0.1
pycodestyle==2.8.0
sqlparse==0.4.2
toml==0.10.2
whitenoise==6.0.0
 